fix(cursor): attest managed hook CLI identity
